Hardset a timezone of Australia/Brisbane.
Make the timezone formatter always return Australia/Brisbane as
currently it is always returning Australia/Melbourne for AEST or +10:00
and this can be different in Brisbane where there is no daylight
savings.

This commit would be a bad idea to apply unless you only have servers in
Brisbane and are encountering the daylight saving bug.