Mod for On Hold status.
Mostly done by NickTheITGuy from OsTicket forum, including a manual merge of PR 3608 which is important if using Teams.

https://github.com/NickTheITGuy/OSTicket-OnHold

Main feature added is correction of counts for open, answered, assigned, and on hold sections.

This involved a change to the ticket counts during the quick staff stats function so that the counter for open tickets for example does not count for on hold tickets.

On hold tickets will be hidden from open, answered, and assigned sections too.