Add in a log event for agent assignment on ticket close.
When a ticket is closed the default behaviour is to change the current
assignment of the ticket to the agent who performs the close. This was
being done without logging to the ost_thread_event table that the
assignment is being changed. This commit adds that log entry. This was
done in order to simplify creation of a statistics query. Particularly
to determine the time between when a ticket was opened and when it was
assigned, as there now will always be at least one assignment logged.
Also added is a check that the currently assigned agent is different to
the agent performing the close as there is no point in updating or
logging if these are the same already.